Phase contrast files moved to "vessels" folder. Pandas and tabulate added
