+++ requirements.txt
@@ -3,8 +3,10 @@ matplotlib>=3.1.3
 nibabel>=3.0.2
 notebook>=6.0.1
 numpy>=1.18.1
+pandas>=1.3.5
 scikit-image>=0.16.2
 scipy>=1.4.1
 SimpleITK>=1.2.4
 scikit-learn>=0.24.2
+tabulate>=0.8.9
 tqdm>=4.48
